refactor(module:input): refactor input
