+++ components/input/demo/autosize-textarea.md
@@ -7,10 +7,11 @@ title:
 
 ## zh-CN
 
-`nzAutosize` 属性适用于 `textarea` 节点，并且只有高度会自动变化。另外 `nzAutosize` 可以设定为一个对象，指定最小行数和最大行数。
+`nzAutosize` 属性适用于 `textarea` 节点，并且只有高度会自动变化。另外 `nzAutosize` 可以设定为一个对象，指定最小行数和最大行数。注意：`nzAutosize` 需要配合 `ngModel` 或者 `ReactiveForm` 使用
 
 ## en-US
 
 `nzAutosize` prop for a `textarea` type of `nz-input` makes the height to automatically adjust based on the content.
 An options object can be provided to `nzAutosize` to specify the minimum and maximum number of lines the textarea will automatically adjust.
+Note: `nzAutosize` must with `ngModel` or `ReactiveForm`.
