+++ components/input/nz-input-group.component.html
@@ -1,10 +1,7 @@
 <span class="ant-input-wrapper ant-input-group" *ngIf="isAddOn">
   <span class="ant-input-group-addon" *ngIf="nzAddOnBefore || nzAddOnBeforeIcon">
     <i nz-icon [ngClass]="nzAddOnBeforeIcon" *ngIf="nzAddOnBeforeIcon"></i>
-    <ng-container *ngIf="isAddOnBeforeString; else addOnBeforeTemplate">{{ nzAddOnBefore }}</ng-container>
-    <ng-template #addOnBeforeTemplate>
-      <ng-template [ngTemplateOutlet]="nzAddOnBefore"></ng-template>
-    </ng-template>
+    <ng-container *nzStringTemplateOutlet="nzAddOnBefore">{{ nzAddOnBefore }}</ng-container>
   </span>
   <ng-template [ngIf]="!isAffix" *ngTemplateOutlet="contentTemplate"></ng-template>
   <span class="ant-input-affix-wrapper" [class.ant-input-affix-wrapper-sm]="isSmall" [class.ant-input-affix-wrapper-lg]="isLarge" *ngIf="isAffix">
@@ -12,10 +9,7 @@
   </span>
   <span class="ant-input-group-addon" *ngIf="nzAddOnAfter || nzAddOnAfterIcon">
     <i nz-icon [ngClass]="nzAddOnAfterIcon" *ngIf="nzAddOnAfterIcon"></i>
-    <ng-container *ngIf="isAddOnAfterString; else addOnAfterTemplate">{{ nzAddOnAfter }}</ng-container>
-    <ng-template #addOnAfterTemplate>
-      <ng-template [ngTemplateOutlet]="nzAddOnAfter"></ng-template>
-    </ng-template>
+    <ng-container *nzStringTemplateOutlet="nzAddOnAfter">{{ nzAddOnAfter }}</ng-container>
   </span>
 </span>
 <ng-container *ngIf="isAffix && !isAddOn">
@@ -25,18 +19,12 @@
   <span class="ant-input-prefix" *ngIf="nzPrefix || nzPrefixIcon">
     <!-- TODO: should have a class to set its color, cc: antd-->
     <i nz-icon [ngClass]="nzPrefixIcon" *ngIf="nzPrefixIcon" style="color: rgba(0, 0, 0, 0.25)"></i>
-    <ng-container *ngIf="isPrefixString; else prefixTemplate">{{ nzPrefix }}</ng-container>
-    <ng-template #prefixTemplate>
-      <ng-template [ngTemplateOutlet]="nzPrefix"></ng-template>
-    </ng-template>
+    <ng-container *nzStringTemplateOutlet="nzPrefix">{{ nzPrefix }}</ng-container>
   </span>
   <ng-template *ngTemplateOutlet="contentTemplate"></ng-template>
   <span class="ant-input-suffix" *ngIf="nzSuffix || nzSuffixIcon">
     <i nz-icon [ngClass]="nzSuffixIcon" *ngIf="nzSuffixIcon"></i>
-    <ng-container *ngIf="isSuffixString; else suffixTemplate">{{ nzSuffix }}</ng-container>
-    <ng-template #suffixTemplate>
-      <ng-template [ngTemplateOutlet]="nzSuffix"></ng-template>
-    </ng-template>
+    <ng-container *nzStringTemplateOutlet="nzSuffix">{{ nzSuffix }}</ng-container>
   </span>
 </ng-template>
 <ng-template [ngIf]="isGroup" *ngTemplateOutlet="contentTemplate"></ng-template>
